Move computeAndSaveDiff into its own file
