+++ packages/cli/src/commands/replay/replay.command.ts
@@ -1,9 +1,6 @@
 import { mkdir, mkdtemp, writeFile } from "fs/promises";
 import { join } from "path";
-import {
-  ScreenshotAssertionsEnabledOptions,
-  ScreenshotDiffResult,
-} from "@alwaysmeticulous/api";
+import { ScreenshotDiffResult } from "@alwaysmeticulous/api";
 import {
   GeneratedBy,
   getMeticulousLocalDataDir,
@@ -20,7 +17,6 @@ import { AxiosInstance } from "axios";
 import log from "loglevel";
 import { DateTime } from "luxon";
 import { createClient } from "../../api/client";
-import { createReplayDiff } from "../../api/replay-diff.api";
 import {
   createReplay,
   getReplayCommandId,
@@ -42,12 +38,6 @@ import {
 } from "../../command-utils/common-types";
 import { sanitizeFilename } from "../../local-data/local-data.utils";
 import { loadReplayEventsDependencies } from "../../local-data/replay-assets";
-import {
-  getOrFetchReplay,
-  getOrFetchReplayArchive,
-  getReplayDir,
-  getScreenshotsDir,
-} from "../../local-data/replays";
 import { serveAssetsFromSimulation } from "../../local-data/serve-assets-from-simulation";
 import {
   getOrFetchRecordedSession,
@@ -56,11 +46,8 @@ import {
 import { getCommitSha } from "../../utils/commit-sha.utils";
 import { addTestCase } from "../../utils/config.utils";
 import { getMeticulousVersion } from "../../utils/version.utils";
-import {
-  checkScreenshotDiffResult,
-  diffScreenshots,
-  ScreenshotDiffError,
-} from "../screenshot-diff/screenshot-diff.command";
+import { ScreenshotDiffError } from "../screenshot-diff/screenshot-diff.command";
+import { computeAndSaveDiff } from "./utils/compute-and-save-diff";
 
 export interface ReplayOptions extends AdditionalReplayOptions {
   replayTarget: ReplayTarget;
@@ -250,70 +237,6 @@ export const replayCommandHandler = async ({
   } finally {
     await deleteArchive(archivePath);
   }
-};
-
-interface ComputeAndSaveDiffOptions {
-  client: AxiosInstance;
-  testRunId: string | null;
-  baseReplayId: string;
-  headReplayId: string;
-  tempDir: string;
-  screenshottingOptions: ScreenshotAssertionsEnabledOptions;
-}
-
-const computeAndSaveDiff = async ({
-  client,
-  baseReplayId,
-  tempDir,
-  headReplayId,
-  screenshottingOptions,
-  testRunId,
-}: ComputeAndSaveDiffOptions): Promise<{
-  screenshotDiffResults: ScreenshotDiffResult[];
-  screenshotDiffError: ScreenshotDiffError | null;
-}> => {
-  const logger = log.getLogger(METICULOUS_LOGGER_NAME);
-
-  logger.info(`Diffing screenshots against replay ${baseReplayId}`);
-
-  await getOrFetchReplay(client, baseReplayId);
-  await getOrFetchReplayArchive(client, baseReplayId);
-
-  const baseReplayScreenshotsDir = getScreenshotsDir(
-    getReplayDir(baseReplayId)
-  );
-  const headReplayScreenshotsDir = getScreenshotsDir(tempDir);
-
-  const screenshotDiffResults = await diffScreenshots({
-    client,
-    baseReplayId,
-    headReplayId,
-    baseScreenshotsDir: baseReplayScreenshotsDir,
-    headScreenshotsDir: headReplayScreenshotsDir,
-    diffOptions: screenshottingOptions.diffOptions,
-  });
-
-  const replayDiff = await createReplayDiff({
-    client,
-    headReplayId,
-    baseReplayId,
-    testRunId,
-    data: {
-      screenshotAssertionsOptions: screenshottingOptions,
-      screenshotDiffResults,
-    },
-  });
-
-  logger.debug(replayDiff);
-
-  const screenshotDiffError = checkScreenshotDiffResult({
-    baseReplayId,
-    headReplayId,
-    results: screenshotDiffResults,
-    diffOptions: screenshottingOptions.diffOptions,
-  });
-
-  return { screenshotDiffResults, screenshotDiffError };
 };
 
 const serveOrGetAppUrl = async (
